# swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:2222/"
        }
    ],
    "paths": {
...
        "/api/entry/": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Entry"
                                                    }
                                                }
                                            },
                                            "required": ["entries"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Entry"
                                                    }
                                                }
                                            },
                                            "required": ["entries"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "New entry added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "facilities": {
                                        "example": "any"
                                    },
                                    "occupierId": {
                                        "example": "any"
                                    },
                                    "authorizationId": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entry/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/entry/release/{id}": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entry released",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "entry": {
                                                    "$ref": "#/components/schemas/Entry"
                                                }
                                            },
                                            "required": ["entry"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facility": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Facility name"
                    },
                    "occupancy": {
                        "type": "array",
                        "items": {
                            "xml": {
                                "name": "facilityoccupancy"
                            },
                            "$ref": "#/components/schemas/FacilityOccupancy"
                        }
                    }
                },
                "required": ["id", "name", "occupancy"],
                "xml": {
                    "name": "Facility"
                }
            },
...
            "FacilityOccupancy": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "isOccupied": {
                        "type": "boolean",
                        "example": true
                    },
                    "facilityId": {
                        "type": "string",
                        "example": "1"
                    },
                    "facility": {
                        "xml": {
                            "name": "facility"
                        },
                        "$ref": "#/components/schemas/Facility"
                    },
                    "createdAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "relatedEntryId": {
                        "type": "string",
                        "example": "2"
                    }
                },
                "required": [
                    "id",
                    "isOccupied",
                    "facilityId",
                    "facility",
                    "createdAt",
                    "relatedEntryId"
                ],
                "xml": {
                    "name": "FacilityOccupancy"
                }
            },
            "Entry": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "authorizationId": {
                        "type": "string",
                        "example": "1"
                    },
                    "occupierId": {
                        "type": "string",
                        "example": "1"
                    },
                    "authorizedById": {
                        "type": "string",
                        "example": "1"
                    },
                    "authorizedBy": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string",
                                "example": "hello@kzakrzewski.pl"
                            }
                        },
                        "required": ["email"]
                    },
                    "occupier": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string",
                                "example": "hello@kzakrzewski.pl"
                            }
                        },
                        "required": ["email"]
                    },
                    "entryAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "releaseAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "relatedOccupancy": {
                        "type": "array",
                        "items": {
                            "xml": {
                                "name": "facilityoccupancy"
                            },
                            "$ref": "#/components/schemas/FacilityOccupancy"
                        }
                    }
                },
                "required": [
                    "id",
                    "occupierId",
                    "authorizedById",
                    "authorizedBy",
                    "occupier",
                    "entryAt",
                    "relatedOccupancy"
                ],
                "xml": {
                    "name": "Entry"
                }
            }
        }
    }
}